In qmos, newer project files will once again work with older versions of qmos (isis 3.4.2 and earlier). Fixes #1478.
git-svn-id: http://subversion.wr.usgs.gov/repos/prog/isis3/trunk@5005 41f8697f-d340-4b68-9986-7bafba869bb8

+++ isis/src/qisis/objs/MosaicSceneWidget/MosaicAreaTool.cpp
@@ -287,7 +287,7 @@ namespace Isis {
       obj += PvlKeyword("Latitude", m_latLineEdit->text());
       obj += PvlKeyword("Longitude", m_lonLineEdit->text());
       obj += PvlKeyword("Area", m_areaLineEdit->text());
-      obj += PvlKeyword("Visible", toString((m_box != NULL)));
+      obj += PvlKeyword("Visible", toString((int)(m_box != NULL)));
     }
 
     return obj;
